% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbits.R
\name{generate_orbits}
\alias{generate_orbits}
\title{Generate concentric circular orbits for "n" planets}
\usage{
generate_orbits(planets, radii = 1:length(planets), num_polygon_points = 1000)
}
\arguments{
\item{planets}{(chr) One or more "planet" names}

\item{radii}{(numeric) By default, this is a set of "n" radii uniformly
spaced apart. You can specify a vector of radii if you want less
uniformly spaced concentric rings.}

\item{num_polygon_points}{(integer) number of points for the circle polygon. Larger == smoother.}
}
\value{
(data.fame)
}
\description{
Given a character vector of "planet" names, generate a
data frame of "n" circle polygons.
}
\examples{
generate_orbits(sol_planets)
}
