% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsolar-package.R
\docType{package}
\name{ggsolar}
\alias{ggsolar}
\alias{ggsolar-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/ggsolar}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/ggsolar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
