% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize-planet-positions.R
\name{randomize_planet_positions}
\alias{randomize_planet_positions}
\title{Generate random planetary positions given a data frame of orbits}
\usage{
randomize_planet_positions(orbits, randomizer = runif)
}
\arguments{
\item{orbits}{(df) output of \code{\link[=generate_orbits]{generate_orbits()}}}

\item{randomizer}{(ƒ) if not the default \code{runif}, a function that can take
a parameter \code{n}, the number of orbits, and return a numeric vector of
values between 0 and 1 that will position the point where you want it
to be on the orbit circle.}
}
\value{
(data.fame) with the \code{planet} name, computed angle, and \code{x}/\code{y} coord of the planet
}
\description{
You can use this after calling \code{\link[=generate_orbits]{generate_orbits()}} to compute a
random position along the orbit circle for each planet
}
\examples{
sol_orbits <- generate_orbits(sol_planets)
randomize_planet_positions(sol_orbits)
}
