% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{sol_planets}
\alias{sol_planets}
\title{Our Solar System}
\format{
An object of class \code{character} of length 9.
}
\usage{
sol_planets
}
\description{
Character vector of the planet names of Sol
}
\note{
To hades with "science"! Pluto is 100\%\% a bona fine planet!
}
\keyword{datasets}
