% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-enhance-solar.R
\name{theme_enhance_solar}
\alias{theme_enhance_solar}
\title{Provides basic cleanup defaults for solar system plots}
\usage{
theme_enhance_solar(bg = "black", fg = "white")
}
\arguments{
\item{bg, fg}{background/foreground colors}
}
\description{
Provides basic cleanup defaults for solar system plots
}
