% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsolar-package.R
\docType{package}
\name{ggsolar}
\alias{ggsolar}
\alias{ggsolar-package}
\title{Generate Solar System Plots}
\description{
Generate "solar system" plots using everyone's favorite plotting package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/ggsolar}
  \item Report bugs at \url{https://github.com/hrbrmstr/ggsolar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
