% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsolar-package.R
\docType{package}
\name{ggsolar}
\alias{ggsolar}
\alias{ggsolar-package}
\title{Generate Solar System Plots}
\description{
Tools are provided to generate a dynamic list of convincing celestial object names
using a small Markov chain model, construct orbital radii of uniform or specified/natural
distances, place celestial objects randomly on the orbit circles using built-in or custom
placement functions. Additional capabilities include the ability to output plots from a core
routine, generate a plotting template to further customize, and place labels next to celestial
object. Some theme elements are also provided to speed up clean plot output generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/ggsolar}
  \item Report bugs at \url{https://github.com/hrbrmstr/ggsolar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
