% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize-planet-positions.R
\name{randomize_planet_positions}
\alias{randomize_planet_positions}
\title{Generate random planetary positions given a data frame of orbits}
\usage{
randomize_planet_positions(orbits, randomizer = default_random_index)
}
\arguments{
\item{orbits}{(df) output of \code{\link[=generate_orbits]{generate_orbits()}}}

\item{randomizer}{(ƒ) a function name that takes as parameters
\code{x} and \code{y} vectors of coordinates (from an orbit polygon)
and a \code{planet} name and returns a data frame of \code{x}, \code{y},
\code{planet} with \code{x} and \code{y} being a single random point from
the \code{planet}'s orbit. By default, this is just (ultimate) a
call to \code{\link[=sample]{sample()}}.}
}
\value{
(data.fame) with the \code{planet} name and \code{x}/\code{y} coords
}
\description{
You can use this after calling \code{\link[=generate_orbits]{generate_orbits()}} to compute a
random position along the orbit circle for each planet
}
\examples{
sol_orbits <- generate_orbits(sol_planets)
randomize_planet_positions(sol_orbits)
}
